OpenAPI & UI: Clean code & support removing additional properties
