+++ packages/ui/src/components/api/playground.tsx
@@ -1,51 +1,23 @@
 import {
-  createContext,
   type HTMLAttributes,
-  type MutableRefObject,
-  type ReactNode,
-  useCallback,
-  useContext,
   useEffect,
   useMemo,
   useRef,
   useState,
 } from 'react';
-import { useFieldArray, useForm, useFormContext } from 'react-hook-form';
-import { PlusIcon, Trash2 } from 'lucide-react';
+import { useForm } from 'react-hook-form';
 import useSWRImmutable from 'swr/immutable';
-import type {
-  APIPlaygroundProps,
-  ReferenceSchema,
-  RequestSchema,
-} from 'fumadocs-openapi';
+import type { APIPlaygroundProps } from 'fumadocs-openapi';
 import { useApiContext } from '@/contexts/api';
-import {
-  Form,
-  FormControl,
-  FormDescription,
-  FormField,
-  FormItem,
-  FormLabel,
-  labelVariants,
-} from '@/components/ui/form';
-import { Input } from '@/components/ui/input';
+import { Form } from '@/components/ui/form';
 import { cn } from '@/utils/cn';
 import { Accordion, Accordions } from '@/components/accordion';
 import * as Base from '@/components/codeblock';
 import { createBodyFromValue, getStatusInfo } from '@/components/api/fetcher';
 import { buttonVariants } from '@/theme/variants';
-import {
-  getDefaultValue,
-  getDefaultValues,
-  resolve,
-} from '@/components/api/shared';
-import {
-  Select,
-  SelectContent,
-  SelectItem,
-  SelectTrigger,
-  SelectValue,
-} from '@/components/ui/select';
+import { getDefaultValue, getDefaultValues } from '@/components/api/shared';
+import { InputField, ObjectInput } from '@/components/api/inputs';
+import { type DynamicField, SchemaContext } from './context';
 
 interface APIPlaygroundFormData {
   authorization?: string | undefined;
@@ -53,29 +25,6 @@ interface APIPlaygroundFormData {
   query?: Record<string, string>;
   header?: Record<string, string>;
   body?: Record<string, unknown>;
-}
-
-export type DynamicField =
-  | {
-      type: 'object';
-      properties: string[];
-    }
-  | {
-      type: 'field';
-      schema: RequestSchema | ReferenceSchema;
-    };
-
-interface SchemaContextType {
-  references: Record<string, RequestSchema>;
-  dynamic: MutableRefObject<Map<string, DynamicField>>;
-}
-
-const SchemaContext = createContext<SchemaContextType | undefined>(undefined);
-
-function useSchemaContext(): SchemaContextType {
-  const ctx = useContext(SchemaContext);
-  if (!ctx) throw new Error('Missing provider');
-  return ctx;
 }
 
 export type CodeBlockProps = HTMLAttributes<HTMLPreElement> & {
@@ -271,11 +220,11 @@ export function APIPlayground({
 
             {body ? (
               <Accordion title="Body">
-                {renderExtracted({
-                  field: body,
-                  fieldName: 'body',
-                  label: 'Body',
-                })}
+                {body.type === 'object' ? (
+                  <ObjectInput field={body} fieldName="body" />
+                ) : (
+                  <InputField name="Body" field={body} fieldName="body" />
+                )}
               </Accordion>
             ) : null}
           </Accordions>
@@ -300,529 +249,5 @@ export function APIPlayground({
         </form>
       </SchemaContext.Provider>
     </Form>
-  );
-}
-
-interface RenderOptions {
-  field: RequestSchema;
-  fieldName: string;
-  className?: string;
-}
-
-function renderInner({ field, ...props }: RenderOptions): React.ReactNode {
-  if (field.type === 'object')
-    return (
-      <ObjectInput
-        field={field}
-        {...props}
-        className={cn('rounded-lg border p-3', props.className)}
-      />
-    );
-  if (field.type === 'switcher') return <Switcher field={field} {...props} />;
-  if (field.type === 'array')
-    return (
-      <ArrayInput
-        field={field}
-        {...props}
-        className={cn('rounded-lg border p-3', props.className)}
-      />
-    );
-  if (field.type === 'null') return null;
-
-  return <NormalInput field={field} {...props} />;
-}
-
-function renderExtracted({
-  field,
-  fieldName,
-  label,
-}: {
-  field: RequestSchema;
-  fieldName: string;
-  label: string;
-}): React.ReactNode {
-  if (field.type === 'object')
-    return <ObjectInput field={field} fieldName={fieldName} />;
-
-  return <InputField name={label} field={field} fieldName={fieldName} />;
-}
-
-interface InputProps<Type> extends HTMLAttributes<HTMLDivElement> {
-  field: Extract<RequestSchema, { type: Type }>;
-  fieldName: string;
-}
-
-function InputContainer(
-  props: {
-    name?: ReactNode;
-    required: boolean;
-    type?: string;
-    description?: string;
-  } & HTMLAttributes<HTMLDivElement>,
-): React.ReactElement {
-  return (
-    <div {...props} className={cn('flex flex-col gap-2', props.className)}>
-      <div className={cn(labelVariants(), 'inline-flex items-center gap-1')}>
-        {props.name}
-        {props.required ? <span className="text-red-500">*</span> : null}
-        {props.type ? (
-          <code className="ms-auto text-xs text-muted-foreground">
-            {props.type}
-          </code>
-        ) : null}
-      </div>
-      <p className="text-xs">{props.description}</p>
-      {props.children}
-    </div>
-  );
-}
-
-function ObjectInput({
-  field,
-  fieldName,
-  ...props
-}: InputProps<'object'>): React.ReactElement {
-  const { references } = useSchemaContext();
-
-  return (
-    <div {...props} className={cn('flex flex-col gap-4', props.className)}>
-      {Object.entries(field.properties).map(([key, child]) => (
-        <InputField
-          key={key}
-          name={key}
-          field={resolve(child, references)}
-          fieldName={`${fieldName}.${key}`}
-        />
-      ))}
-      {field.additionalProperties ? (
-        <AdditionalProperties
-          fieldName={fieldName}
-          type={field.additionalProperties}
-        />
-      ) : null}
-    </div>
-  );
-}
-
-function AdditionalProperties({
-  fieldName,
-  type,
-}: {
-  fieldName: string;
-  type: boolean | string;
-}): React.ReactElement {
-  const { references, dynamic } = useSchemaContext();
-  const [nextName, setNextName] = useState('');
-  const [properties, setProperties] = useState<string[]>(() => {
-    const d = dynamic.current.get(fieldName);
-    if (d?.type === 'object') return d.properties;
-
-    return [];
-  });
-
-  useEffect(() => {
-    dynamic.current.set(fieldName, { type: 'object', properties });
-  }, [properties]);
-
-  const onAppend = useCallback(() => {
-    if (nextName.length === 0) return;
-    setProperties((p) => (p.includes(nextName) ? p : [...p, nextName]));
-    setNextName('');
-  }, [nextName]);
-
-  const types =
-    typeof type === 'string'
-      ? resolveDynamicField(references[type], references)
-      : undefined;
-
-  return (
-    <>
-      {properties.map((item) => (
-        <DynamicField
-          key={item}
-          label={item}
-          types={types}
-          fieldName={`${fieldName}.${item}`}
-        />
-      ))}
-      <div className="flex flex-row gap-1">
-        <Input
-          value={nextName}
-          placeholder="Enter Property Name"
-          onChange={useCallback((e: React.ChangeEvent<HTMLInputElement>) => {
-            setNextName(e.target.value);
-          }, [])}
-          onKeyDown={useCallback(
-            (e: React.KeyboardEvent<HTMLInputElement>) => {
-              if (e.key === 'Enter') {
-                onAppend();
-                e.preventDefault();
-              }
-            },
-            [onAppend],
-          )}
-        />
-        <button
-          type="button"
-          className={cn(buttonVariants({ color: 'secondary' }))}
-          onClick={onAppend}
-        >
-          New
-        </button>
-      </div>
-    </>
-  );
-}
-
-function resolveDynamicField(
-  schema: RequestSchema,
-  references: Record<string, RequestSchema>,
-): Record<string, RequestSchema> {
-  if (schema.type !== 'switcher') return { [schema.type]: schema };
-
-  return Object.fromEntries(
-    Object.entries(schema.items).map(([key, value]) => [
-      key,
-      resolve(value, references),
-    ]),
-  );
-}
-
-function DynamicField({
-  fieldName,
-  label,
-  types = {
-    string: {
-      type: 'string',
-      isRequired: false,
-      defaultValue: '',
-    },
-    boolean: {
-      type: 'boolean',
-      isRequired: false,
-      defaultValue: '',
-    },
-    number: {
-      type: 'number',
-      isRequired: false,
-      defaultValue: '',
-    },
-  },
-  className,
-}: {
-  fieldName: string;
-  label: string;
-  /**
-   * Available types, fallback to any
-   */
-  types?: Record<string, RequestSchema>;
-  className?: string;
-}): React.ReactElement {
-  const { dynamic } = useSchemaContext();
-  const typeNames = Object.keys(types);
-  const [value, setValue] = useState<string>(() => {
-    const d = dynamic.current.get(fieldName);
-
-    if (d?.type === 'field') {
-      return typeNames.find((name) => types[name] === d.schema) ?? typeNames[0];
-    }
-
-    return typeNames[0];
-  });
-
-  useEffect(() => {
-    if (!value) return;
-
-    dynamic.current.set(fieldName, {
-      type: 'field',
-      schema: types[value],
-    });
-  }, [value, fieldName, types, dynamic]);
-
-  return (
-    <InputContainer
-      name={
-        <>
-          {label}
-          <Select value={value} onValueChange={setValue}>
-            <SelectTrigger className="ms-auto h-auto gap-1 p-1 text-xs">
-              <SelectValue />
-            </SelectTrigger>
-            <SelectContent>
-              {typeNames.map((item) => (
-                <SelectItem key={item} value={item}>
-                  {item}
-                </SelectItem>
-              ))}
-            </SelectContent>
-          </Select>
-        </>
-      }
-      className={className}
-      required={false}
-    >
-      {renderInner({
-        field: types[value],
-        fieldName,
-      })}
-    </InputContainer>
-  );
-}
-
-function Switcher({
-  field,
-  fieldName,
-  className,
-}: InputProps<'switcher'>): React.ReactElement {
-  const { references, dynamic } = useSchemaContext();
-  const items = Object.keys(field.items);
-  const [value, setValue] = useState<string>(() => {
-    const d = dynamic.current.get(fieldName);
-
-    if (d?.type === 'field') {
-      // schemas are passed from server components, they shouldn't be re-constructed
-      const cached = items.find((item) => d.schema === field.items[item]);
-
-      if (cached) return cached;
-    }
-
-    return items[0];
-  });
-
-  useEffect(() => {
-    if (!value) return;
-
-    dynamic.current.set(fieldName, {
-      type: 'field',
-      schema: field.items[value],
-    });
-  }, [value, fieldName, field, dynamic]);
-
-  return (
-    <div className={className}>
-      <Select value={value} onValueChange={setValue}>
-        <SelectTrigger>
-          <SelectValue />
-        </SelectTrigger>
-        <SelectContent>
-          {items.map((item) => (
-            <SelectItem key={item} value={item}>
-              {item}
-            </SelectItem>
-          ))}
-        </SelectContent>
-      </Select>
-      {renderExtracted({
-        field: resolve(field.items[value], references),
-        fieldName,
-        label: 'Value',
-      })}
-    </div>
-  );
-}
-
-function InputField({
-  field,
-  name,
-  fieldName,
-  ...props
-}: {
-  name?: string;
-  field: RequestSchema;
-  fieldName: string;
-  className?: string;
-}): React.ReactNode {
-  const { references } = useSchemaContext();
-
-  if (field.type === 'object') {
-    return (
-      <InputContainer
-        name={name}
-        required={field.isRequired}
-        type={field.type}
-        description={field.description}
-        {...props}
-      >
-        <ObjectInput
-          field={field}
-          fieldName={fieldName}
-          className="rounded-lg border bg-accent/30 p-3"
-        />
-      </InputContainer>
-    );
-  }
-
-  if (field.type === 'array') {
-    return (
-      <InputContainer
-        name={name}
-        required={field.isRequired}
-        description={field.description ?? references[field.items].description}
-        type={`array<${references[field.items].type}>`}
-        {...props}
-      >
-        <ArrayInput
-          fieldName={fieldName}
-          field={field}
-          className="rounded-lg border bg-background p-3"
-        />
-      </InputContainer>
-    );
-  }
-
-  if (field.type === 'switcher') {
-    return (
-      <InputContainer
-        name={name}
-        description={field.description}
-        required={field.isRequired}
-        {...props}
-      >
-        <Switcher field={field} fieldName={fieldName} />
-      </InputContainer>
-    );
-  }
-
-  if (field.type === 'null') return null;
-
-  return (
-    <NormalInput
-      field={field}
-      fieldName={fieldName}
-      header={
-        <>
-          <FormLabel className="inline-flex gap-1">
-            {name}
-            {field.isRequired ? <span className="text-red-500">*</span> : null}
-            <code className="ms-auto text-xs text-muted-foreground">
-              {field.type}
-            </code>
-          </FormLabel>
-          <FormDescription className="text-xs">
-            {field.description}
-          </FormDescription>
-        </>
-      }
-      {...props}
-    />
-  );
-}
-
-function NormalInput({
-  fieldName,
-  header,
-  field,
-  ...props
-}: InputProps<'string' | 'boolean' | 'number'> & {
-  header?: React.ReactNode;
-}): React.ReactElement {
-  const { control } = useFormContext();
-
-  if (field.type === 'boolean') {
-    return (
-      <FormField
-        control={control}
-        name={fieldName}
-        render={({ field: { value, onChange, ...restField } }) => (
-          <FormItem {...props}>
-            {header}
-            <Select
-              value={value as string}
-              onValueChange={onChange}
-              disabled={restField.disabled}
-            >
-              <FormControl>
-                <SelectTrigger {...restField}>
-                  <SelectValue />
-                </SelectTrigger>
-              </FormControl>
-              <SelectContent>
-                <SelectItem value="true">True</SelectItem>
-                <SelectItem value="false">False</SelectItem>
-                {field.isRequired ? null : (
-                  <SelectItem value="null">Null</SelectItem>
-                )}
-              </SelectContent>
-            </Select>
-          </FormItem>
-        )}
-      />
-    );
-  }
-
-  return (
-    <FormField
-      control={control}
-      name={fieldName}
-      render={({ field: { value, ...restField } }) => (
-        <FormItem {...props}>
-          {header}
-          <FormControl>
-            <Input
-              placeholder="Enter value"
-              type={field.type === 'string' ? 'text' : 'number'}
-              value={value as string}
-              {...restField}
-            />
-          </FormControl>
-        </FormItem>
-      )}
-    />
-  );
-}
-
-function ArrayInput({
-  fieldName,
-  field,
-  ...props
-}: InputProps<'array'>): React.ReactElement {
-  const { references } = useSchemaContext();
-  const { fields, append, remove } = useFieldArray({
-    name: fieldName,
-  });
-  const items = references[field.items];
-
-  const handleAppend = useCallback(() => {
-    append(getDefaultValue(items, references));
-  }, [append, references, items]);
-
-  return (
-    <div {...props} className={cn('flex flex-col gap-4', props.className)}>
-      {fields.map((item, index) => (
-        <div key={item.id} className="relative">
-          {renderInner({
-            field: items,
-            fieldName: `${fieldName}.${String(index)}`,
-            className: 'flex-1',
-          })}
-          <button
-            type="button"
-            aria-label="Remove Item"
-            className={cn(
-              buttonVariants({
-                color: 'outline',
-                size: 'sm',
-                className: 'absolute bg-background -top-2 -end-2',
-              }),
-            )}
-            onClick={() => {
-              remove(index);
-            }}
-          >
-            <Trash2 className="size-4" />
-          </button>
-        </div>
-      ))}
-      <button
-        type="button"
-        className={cn(
-          buttonVariants({ color: 'secondary', className: 'gap-1.5' }),
-        )}
-        onClick={handleAppend}
-      >
-        <PlusIcon className="size-4" />
-        New Item
-      </button>
-    </div>
   );
 }
